% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bootstrap_table}
\alias{bootstrap_table}
\title{Create a bootstrap html table}
\usage{
bootstrap_table(.rownames, .values)
}
\arguments{
\item{.rownames}{vector of length three}

\item{.values}{vector of length three}
}
\value{
html code for a 2x3 table
}
\description{
Similar to knitr::kable and kableExtra::kable_styling without the dependencies.
}
\examples{
my_names <- c('n sequences', 'n unique seqences', 'n periods')
my_values <- c(50, 20, 10)
bootstrap_table(my_names, my_values)
}
