% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cluster_labels}
\alias{cluster_labels}
\title{Label clusters consistently}
\usage{
cluster_labels(.model, k)
}
\arguments{
\item{.model}{an hclust model}

\item{k}{the number of clusters}
}
\value{
a factor
}
\description{
Returns cluster labels that match the clusters in sequenchr::plot_dendrogram branches left-to-right.
}
\examples{
library(TraMineR)
data(mvad)
seqstatl(mvad[, 17:86])
mvad.alphabet <- c("employment", "FE", "HE", "joblessness", "school",
                   "training")
mvad.labels <- c("employment", "further education", "higher education",
                 "joblessness", "school", "training")
mvad.seq <- seqdef(mvad, 17:86, alphabet = mvad.alphabet, # states = mvad.scodes,
                   labels = mvad.labels, xtstep = 6)
dist_matrix <- TraMineR::seqdist(seqdata = mvad.seq, method = "DHD")
cluster_model <- hclust(d = as.dist(dist_matrix), method = 'ward.D2')

cluster_labels(cluster_model, k = 5)
}
