% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cluster_stats}
\alias{cluster_stats}
\title{Compute CH index and silhouette width}
\usage{
cluster_stats(dist_matrix, cluster_model, k_min, k_max)
}
\arguments{
\item{dist_matrix}{a distance matrix}

\item{cluster_model}{a clustering model such as the output from fastcluster::hclust}

\item{k_min}{the minimum number of clusters to test}

\item{k_max}{the maximum number of clusters to test}
}
\value{
tibble
}
\description{
Computes the raw and normalized Calinski-Harabasz index and silhouette width for various number of clusters.
}
\examples{
library(TraMineR)
data(mvad)
seqstatl(mvad[, 17:86])
mvad.alphabet <- c("employment", "FE", "HE", "joblessness", "school",
                   "training")
mvad.labels <- c("employment", "further education", "higher education",
                 "joblessness", "school", "training")
mvad.seq <- seqdef(mvad, 17:86, alphabet = mvad.alphabet, # states = mvad.scodes,
                   labels = mvad.labels, xtstep = 6)
dist_matrix <- TraMineR::seqdist(seqdata = mvad.seq, method = "DHD")
cluster_model <- fastcluster::hclust(d = as.dist(dist_matrix), method = 'ward.D2')

cluster_stats(
 dist_matrix = as.dist(dist_matrix),
 cluster_model = cluster_model,
 k_min = 2,
 k_max = 5
)

}
