% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_dendrogram}
\alias{plot_dendrogram}
\title{Plot a dendrogram colored by cluster}
\usage{
plot_dendrogram(cluster_model, k, h = 100)
}
\arguments{
\item{cluster_model}{a clustering model such as the output from fastcluster::hclust}

\item{k}{the number of clusters}

\item{h}{the minimum height to plot the segments. A lower height results in decreased performance}
}
\value{
ggplot object
}
\description{
Plots a dedrogram where the colors of the segments represent cluster membership
}
\examples{
library(TraMineR)
data(mvad)
seqstatl(mvad[, 17:86])
mvad.alphabet <- c("employment", "FE", "HE", "joblessness", "school",
                   "training")
mvad.labels <- c("employment", "further education", "higher education",
                 "joblessness", "school", "training")
mvad.seq <- seqdef(mvad, 17:86, alphabet = mvad.alphabet, # states = mvad.scodes,
                   labels = mvad.labels, xtstep = 6)
dist_matrix <- TraMineR::seqdist(seqdata = mvad.seq, method = "DHD")
cluster_model <- fastcluster::hclust(d = as.dist(dist_matrix), method = 'ward.D2')
plot_dendrogram(cluster_model, 5)
}
