% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_legend}
\alias{plot_legend}
\title{Plot the legend}
\usage{
plot_legend(color_mapping)
}
\arguments{
\item{color_mapping}{a list of named colors}
}
\value{
ggplot object
}
\description{
Plots just the legend given a list of named oclors
}
\examples{
library(TraMineR)
data(mvad)
seqstatl(mvad[, 17:86])
mvad.alphabet <- c("employment", "FE", "HE", "joblessness", "school",
                   "training")
mvad.labels <- c("employment", "further education", "higher education",
                 "joblessness", "school", "training")
mvad.seq <- seqdef(mvad, 17:86, alphabet = mvad.alphabet, # states = mvad.scodes,
                   labels = mvad.labels, xtstep = 6)
color_mapping <- viridis::viridis_pal()(length(alphabet(mvad.seq)))
names(color_mapping) <- alphabet(mvad.seq)
plot_legend(color_mapping)
}
