% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_sequence_index}
\alias{plot_sequence_index}
\title{Generates a sequence index plot}
\usage{
plot_sequence_index(
  seq_def_tidy,
  color_mapping = NULL,
  cluster_labels = NULL,
  n_col_facets = 1
)
}
\arguments{
\item{seq_def_tidy}{a tidy tibble generated from sequenchr::tidy_sequence_data}

\item{color_mapping}{optional. A list of named colors where the names match the alphabet of the original sequence data. Useful for ensuring consistent legends across plots.}

\item{cluster_labels}{optional. A vector of cluster assignments}

\item{n_col_facets}{optional. If cluster_labels is provided then the number of facet columns}
}
\value{
ggplot object
}
\description{
Generates a sequence index plot
}
\examples{
library(TraMineR)
data(mvad)
seqstatl(mvad[, 17:86])
mvad.alphabet <- c("employment", "FE", "HE", "joblessness", "school",
                   "training")
mvad.labels <- c("employment", "further education", "higher education",
                 "joblessness", "school", "training")
mvad.seq <- seqdef(mvad, 17:86, alphabet = mvad.alphabet,
                   labels = mvad.labels, xtstep = 6)
seq_def_tidy <- tidy_sequence_data(mvad.seq)
color_mapping <- viridis::viridis_pal()(length(alphabet(mvad.seq)))
names(color_mapping) <- alphabet(mvad.seq)
plot_sequence_index(seq_def_tidy, color_mapping)

dist_matrix <- TraMineR::seqdist(seqdata = mvad.seq, method = "DHD")
cluster_model <- hclust(d = as.dist(dist_matrix), method = 'ward.D2')
cluster_labels <- stats::cutree(cluster_model, k = 5)
plot_sequence_index(seq_def_tidy, color_mapping, cluster_labels = cluster_labels)
}
