% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_transition_matrix}
\alias{plot_transition_matrix}
\title{Plot a transition matrix}
\usage{
plot_transition_matrix(transition_matrix)
}
\arguments{
\item{transition_matrix}{a transition matrix}
}
\value{
ggplot object
}
\description{
Plots a 'heatmap' of a transition matrix
}
\examples{
#TODO
}
