% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{shannon_entropy}
\alias{shannon_entropy}
\title{Shannon entropy of a vector}
\usage{
shannon_entropy(x)
}
\arguments{
\item{x}{a categorical vector}
}
\value{
numeric
}
\description{
Calculates the Shannon entropy of a categorical vector
}
\examples{
shannon_entropy(c('A', 'A', 'B', 'C', 'A', 'C'))
}
