% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tidy_sequence_data}
\alias{tidy_sequence_data}
\title{Convert output of TraMineR::seqdef to tidy dataframe}
\usage{
tidy_sequence_data(sequence_data)
}
\arguments{
\item{sequence_data}{}
}
\value{
tibble
}
\description{
Converts the output matrix of TraMineR::seqdef into a tidy tibble
}
\examples{
library(TraMineR)
data(mvad)
seqstatl(mvad[, 17:86])
mvad.alphabet <- c("employment", "FE", "HE", "joblessness", "school",
                   "training")
mvad.labels <- c("employment", "further education", "higher education",
                 "joblessness", "school", "training")
mvad.seq <- seqdef(mvad, 17:86, alphabet = mvad.alphabet, # states = mvad.scodes,
                   labels = mvad.labels, xtstep = 6)
tidy_sequence_data(mvad.seq)
}
