% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_sequenchr.R
\name{launch_sequenchr}
\alias{launch_sequenchr}
\title{Launches the sequenchr Shiny app}
\usage{
launch_sequenchr(MonteCarlo = 10)
}
\arguments{
\item{sequence_data}{an object created from TraMineR::seqdef}

\item{covariates_data}{optional. a dataframe with the same number of rows as sequence_data}
}
\value{

}
\description{
Launches the sequenchr app in the default web browser. Optional covariates_data produces plots of the marginal distributions.
}
